backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
psycopg2-binary==2.9.9
celery==5.4.0
django-celery-beat==2.6.0
redis==5.1.1
pyotp==2.9.0
bcrypt==4.2.0
scikit-learn==1.5.2
gunicorn==23.0.0
numpy==1.26.4
joblib==1.4.2
reportlab==4.2.5
requests==2.32.3
